Admin Management Functionality
Added the ability for an admin to get information regarding all users, and added some functionality for deleting users as well. Additionally, when you click on a user in the manage users component, it will navigate them to a user details component. Here a admin can see all the details for the user they are viewing.