Added Custom Input for Instruments and Improving Input/Selection Accessibility.
When selecting "Other" under the instrument dropdown. It now provides an input to enter the custom name. This did create a problem with the student model since we were originally validating instrument selection at the model level and the adduserform. To remedy this I had to remove the required requirement for instrument and I also included a new customInstrument field.

Aside from that, I also updated the labels I added because React framework prefers when you add htmlFor= "<whatever the select or input field is >", and then an id which references that input or selection. Supposedly this helps with autofill and other accessibility features.