User Details Form Work
I added the date/time created to the user details, and convert the timestamp from mongodb into a more readable form in the users local time.
Additionally, when fetching the parent details for students and student details for parents, it was fetching the ID's not the names in a readable format. So I made those additions as well, we can now see the names and not their ID's. We already had a method for fetching the children of a parent, but we did not have one for fetching the parent of a child, so I add a getParent method to the studentController and it's route into studentRoutes.