Work on Advanced User View
This will hopefully fulfill the need for a more grid-like view of user types. I developed a AdvancedView button in the manage users component, which navigates to the advanced view. The advanced view will kind of act like the excel spreadsheet window where we will allow users to select the equivalent of tabs representing the different user types. Each one of these tabs is it's own component.

So far I have only built the StudentViewTab, but it works great from what I can see. You can click on the row and navigate to the user details, and even click on their parent's cells and instructors and navigate to their user details.

I also added a feature for them to actually just export the grid of users in the tab directly into an excel document.

There needs to be more robust sorting in the application but this is a great start, and the work done so far can easily be translated to the other user types.